Separar completamente los componentes del resourcepack tab
