+++ package.json
@@ -34,7 +34,7 @@
 	"devDependencies": {
 		"@eslint/js": "^10.0.1",
 		"@sveltejs/adapter-static": "^3.0.10",
-		"@sveltejs/kit": "^2.67.0",
+		"@sveltejs/kit": "^2.68.0",
 		"@sveltejs/vite-plugin-svelte": "^5.1.1",
 		"@tauri-apps/cli": "^2.11.3",
 		"@types/node": "^25.9.4",
@@ -45,6 +45,7 @@
 		"lint-staged": "^17.0.8",
 		"prettier": "^3.8.4",
 		"prettier-plugin-svelte": "^4.1.1",
+		"sass-embedded": "^1.100.0",
 		"svelte": "^5.56.4",
 		"svelte-check": "^4.7.1",
 		"typescript": "~5.6.3",
