+++ src/lib/icons/Cubic.svelte
@@ -1,13 +1,21 @@
 <script lang="ts">
-	let { height, width }: { height: string; width: string } = $props();
+	let {
+		width,
+		height,
+	}: {
+		width?: number | string;
+		height?: number | string;
+	} = $props();
 </script>
 
 <svg
 	version="1.1"
 	xmlns="http://www.w3.org/2000/svg"
-	{width}
-	{height}
 	fill="currentColor"
+	viewBox="0 0 512 512"
+	preserveAspectRatio="xMidYMid meet"
+	width={width ?? "100%"}
+	height={height ?? "100%"}
 >
 	<path
 		d="M0 0 C4.71460177 1.84496716 8.91177367 4.38328336 13.1875 7.0625 C14.40629496 7.79644467 15.62687692 8.52742905 16.84912109 9.25561523 C19.35542912 10.75278253 21.85283763 12.26331897 24.34423828 13.78515625 C29.87503673 17.15244346 35.50511361 20.34703991 41.125 23.5625 C43.27405285 24.8023194 45.42249081 26.04320525 47.5703125 27.28515625 C54.04510022 31.02651656 60.52236343 34.76357483 67 38.5 C69.1666775 39.74998122 71.33334417 40.99998121 73.5 42.25 C75.10875 43.178125 75.10875 43.178125 76.75 44.125 C164.5 94.75 164.5 94.75 167.75415039 96.62744141 C169.90692816 97.86939744 172.05975963 99.1112604 174.21264648 100.35302734 C179.24616915 103.25646913 184.27918048 106.160785 189.31021118 109.06854248 C191.08362608 110.09345403 192.85721646 111.11806123 194.63085938 112.14257812 C196.84995773 113.42452738 199.06873016 114.70704107 201.28710938 115.99023438 C202.29708984 116.57353516 203.30707031 117.15683594 204.34765625 117.7578125 C205.66262085 118.51795654 205.66262085 118.51795654 207.00415039 119.29345703 C210.98472785 121.56093616 214.99539989 123.77522216 219 126 C219 195.96 219 265.92 219 338 C215.05447331 337.01361833 212.24940212 335.66854339 208.6796875 333.75390625 C207.75339035 333.2588987 207.75339035 333.2588987 206.80838013 332.75389099 C205.49435682 332.0506369 204.18146537 331.34526445 202.86962891 330.63793945 C200.93233785 329.59341471 198.99196371 328.5548493 197.05078125 327.51757812 C186.9741107 322.11318822 176.9978055 316.55433639 167 311 C167 262.16 167 213.32 167 163 C165.35 162.34 163.7 161.68 162 161 C160.09969676 160.0098312 158.21529464 158.98881972 156.34375 157.9453125 C155.25006104 157.33695557 154.15637207 156.72859863 153.02954102 156.10180664 C151.85293541 155.44303527 150.67642403 154.78409558 149.5 154.125 C148.27602611 153.44185797 147.0519856 152.75883528 145.82788086 152.07592773 C136.12474193 146.65610245 126.4660489 141.16006982 116.82910156 135.62353516 C111.96368968 132.83142357 107.0918118 130.05062797 102.22067261 127.26852417 C99.22889636 125.55966188 96.23751615 123.85010661 93.24609375 122.140625 C92.03499537 121.44856442 90.82389511 120.75650713 89.61279297 120.06445312 C86.04101562 118.0234375 86.04101562 118.0234375 82.46923828 115.98242188 C81.89143616 115.65225067 81.31363403 115.32207947 80.71832275 114.98190308 C79.55470284 114.3169745 78.39108793 113.65203717 77.22747803 112.98709106 C74.32120324 111.32631175 71.41484121 109.66568547 68.50830078 108.00537109 C63.22846387 104.9891511 57.9492897 101.97179938 52.67330933 98.94883728 C49.72013116 97.25681571 46.76641983 95.56572639 43.8125 93.875 C43.23681274 93.54543304 42.66112549 93.21586609 42.06799316 92.87631226 C32.47126423 87.38501575 22.83676518 81.96503265 13.16723633 76.60302734 C12.20970459 76.07016113 11.25217285 75.53729492 10.265625 74.98828125 C9.36521484 74.4876416 8.46480469 73.98700195 7.53710938 73.47119141 C4.50586763 71.71346789 1.60949473 69.79126614 -1.25488281 67.77416992 C-3.63979584 66.71617339 -4.57657536 67.06836311 -7 68 C-8.88136336 68.96271025 -10.73398423 69.98227038 -12.56640625 71.03515625 C-13.67773926 71.67138916 -14.78907227 72.30762207 -15.93408203 72.96313477 C-17.14368118 73.66270451 -18.35315033 74.36249906 -19.5625 75.0625 C-20.82132824 75.78659185 -22.08028208 76.51046538 -23.33935547 77.23413086 C-32.9364147 82.76123661 -42.46790071 88.39679275 -51.98388672 94.06225586 C-60.75877056 99.28211801 -69.58688741 104.40397461 -78.4375 109.49414062 C-88.66139385 115.37442866 -98.84933435 121.31518138 -109.03222656 127.26611328 C-112.01818929 129.01062688 -115.00501287 130.75365622 -117.9921875 132.49609375 C-118.98484619 133.07512451 -119.97750488 133.65415527 -121.00024414 134.25073242 C-122.98619589 135.40881367 -124.97235831 136.56653371 -126.95874023 137.72387695 C-132.20138042 140.77938756 -137.4383558 143.84411296 -142.66796875 146.921875 C-143.75916016 147.56253906 -144.85035156 148.20320312 -145.97460938 148.86328125 C-148.06740714 150.09224334 -150.15863644 151.32388152 -152.24804688 152.55859375 C-153.6624707 153.38810547 -153.6624707 153.38810547 -155.10546875 154.234375 C-155.93199951 154.72131836 -156.75853027 155.20826172 -157.61010742 155.70996094 C-160.01925051 157.01039123 -162.4401968 158.03498239 -165 159 C-165.33086152 163.55074425 -165.22725496 167.58688611 -164 172 C-160.68620744 175.50037827 -156.74156443 177.35002571 -152.4375 179.4375 C-151.25070398 180.05856187 -150.06664659 180.6848825 -148.88525391 181.31616211 C-146.455684 182.60953269 -144.01622097 183.88021654 -141.56689453 185.13574219 C-136.21568344 187.90217591 -131.05473752 190.99051953 -125.875 194.0625 C-123.89401578 195.22139775 -121.91224375 196.37894988 -119.9296875 197.53515625 C-118.95225586 198.10637207 -117.97482422 198.67758789 -116.96777344 199.26611328 C-111.98009043 202.1801084 -106.98991295 205.08982568 -102 208 C-99.9999882 209.16664643 -97.99998819 210.33331309 -96 211.5 C-95.01 212.0775 -94.02 212.655 -93 213.25 C-82 219.66666667 -71 226.08333333 -60 232.5 C-59.0085498 233.07830566 -58.01709961 233.65661133 -56.99560547 234.25244141 C-55.010112 235.41072684 -53.02475874 236.56925265 -51.03955078 237.72802734 C-45.81876798 240.77498642 -40.59509463 243.81684106 -35.3671875 246.8515625 C-34.28050781 247.48320312 -33.19382813 248.11484375 -32.07421875 248.765625 C-29.99699247 249.97290181 -27.91891564 251.17871681 -25.83984375 252.3828125 C-24.90011719 252.929375 -23.96039063 253.4759375 -22.9921875 254.0390625 C-22.17089355 254.51553223 -21.34959961 254.99200195 -20.50341797 255.48291016 C-19.67729004 255.9835498 -18.85116211 256.48418945 -18 257 C-16.9732485 257.56235226 -15.946497 258.12470451 -14.88863182 258.70409775 C-12 261 -12 261 -11.26171303 263.86603165 C-11.26724379 264.98324474 -11.27277454 266.10045784 -11.2784729 267.25152588 C-11.27022491 268.53597595 -11.26197693 269.82042603 -11.253479 271.14379883 C-11.2747127 272.56494982 -11.29652154 273.98609231 -11.31884766 275.40722656 C-11.32054029 276.90533378 -11.31958973 278.40344597 -11.31619263 279.90155029 C-11.31447008 283.96773519 -11.34976057 288.03285954 -11.39163637 292.0987978 C-11.42912798 296.34817957 -11.43266746 300.59758529 -11.43978882 304.84710693 C-11.45844239 312.89371616 -11.50768601 320.93983666 -11.56788576 328.98623061 C-11.63493525 338.14690221 -11.66793591 347.30759707 -11.69805837 356.46845293 C-11.76078486 375.31257911 -11.86632952 394.15625027 -12 413 C-15.59228612 411.58902773 -18.96689013 410.02906765 -22.3671875 408.19921875 C-23.40544678 407.64113525 -24.44370605 407.08305176 -25.51342773 406.50805664 C-27.17740601 405.6068811 -27.17740601 405.6068811 -28.875 404.6875 C-30.03394775 404.06157959 -31.19289551 403.43565918 -32.38696289 402.79077148 C-41.69130685 397.75407217 -50.90828961 392.56919591 -60.10546875 387.33984375 C-66.46453518 383.72768816 -72.854884 380.17284648 -79.25 376.625 C-80.37623779 376.00004639 -81.50247559 375.37509277 -82.6628418 374.73120117 C-86.10818363 372.82011887 -89.55400274 370.90990018 -93 369 C-106.39448125 361.57627904 -119.78476549 354.14590821 -133.125 346.625 C-134.10952148 346.07118652 -135.09404297 345.51737305 -136.10839844 344.94677734 C-137.00333008 344.44098145 -137.89826172 343.93518555 -138.8203125 343.4140625 C-140.00004639 342.74818115 -140.00004639 342.74818115 -141.20361328 342.06884766 C-143 341 -143 341 -144 340 C-144.26273524 336.16245954 -144.18012513 332.30625721 -144.17578125 328.4609375 C-144.17690666 326.6573024 -144.17690666 326.6573024 -144.17805481 324.81723022 C-144.13429115 316.2052731 -143.73594068 307.65337224 -143.17010117 299.06122589 C-142.91363006 294.98180626 -142.71661601 290.8990103 -142.52075005 286.81628799 C-142.34912001 283.34256459 -142.13646537 279.87977522 -141.84301758 276.4140625 C-140.04247808 261.25006419 -140.04247808 261.25006419 -144.85470581 247.56594849 C-149.74114567 243.28813014 -155.33012362 240.67406049 -161.2807312 238.1809082 C-165.12877375 236.50980272 -168.58959229 234.35952906 -172.13671875 232.13671875 C-173.40571741 231.38187313 -174.6771914 230.63117251 -175.95117188 229.88476562 C-176.57048584 229.52036377 -177.1897998 229.15596191 -177.82788086 228.78051758 C-179.05862728 228.05712888 -180.28942086 227.33382038 -181.52026367 226.6105957 C-182.129104 226.25247803 -182.73794434 225.89436035 -183.36523438 225.52539062 C-183.9910994 225.15912567 -184.61696442 224.79286072 -185.26179504 224.41549683 C-186.5653205 223.63921898 -187.85855222 222.84541812 -189.1414032 222.03543091 C-193.53626804 219.29319652 -197.93746351 217.14228087 -202.7121582 215.12255859 C-219.46602836 207.43755042 -219.46602836 207.43755042 -222.83603287 200.20598984 C-226.26756195 189.86509501 -225.37394859 178.47682746 -224.89276123 167.76080322 C-224.78554699 164.40683204 -224.84680634 161.06412132 -224.90516281 157.70961857 C-225.23697581 133.64432604 -225.23697581 133.64432604 -220.17297363 126.37863159 C-215.81318139 122.27831625 -210.87459417 119.95475913 -205.30709839 117.89117432 C-202.23105234 116.70297478 -199.5996181 115.03147236 -196.8515625 113.22265625 C-195.11025447 112.16781765 -193.36807805 111.11441113 -191.625 110.0625 C-184.29137069 105.61686833 -176.97487374 101.14607773 -169.6875 96.625 C-153.5607058 86.62246022 -137.40610913 76.6655705 -121.1875 66.8125 C-113.14085382 61.92382433 -105.12068801 56.99472322 -97.11938477 52.03222656 C-81.84822567 42.56134249 -66.54505939 33.14266533 -51.1875 23.8125 C-43.58712851 19.19406094 -36.00214861 14.55175031 -28.4375 9.875 C-27.14332153 9.07606323 -27.14332153 9.07606323 -25.82299805 8.26098633 C-25.01386963 7.75978271 -24.20474121 7.2585791 -23.37109375 6.7421875 C-22.60265137 6.27393555 -21.83420898 5.80568359 -21.04248047 5.32324219 C-18.79824579 3.93240757 -18.79824579 3.93240757 -16.22265625 1.6328125 C-10.87393988 -1.0772038 -5.85383388 -0.96080916 0 0 Z "
